requirements.txt
aniso8601>=9.0.1
appdirs>=1.4.4
astroid>=2.5.2
bcrypt>=3.2.0
certifi>=2020.12.5
cffi>=1.14.5
chardet>=4.0.0
click>=7.1.2
colorama>=0.4.4
DateTime>=4.3
distlib>=0.3.1
filelock>=3.0.12
Flask>=1.1.2
Flask-Bcrypt>=0.7.1
Flask-Cors>=3.0.10
Flask-JWT>=0.3.2
Flask-PyMongo>=2.3.0
Flask-RESTful>=0.3.8
idna>=2.10
isort>=5.8.0
itsdangerous>=1.1.0
Jinja2>=2.11.3
lazy-object-proxy>=1.6.0
MarkupSafe>=1.1.1
mccabe>=0.6.1
pip>=20.2.3
pycparser>=2.20
pylint>=2.7.4
pymongo>=3.11.3
pytz>=2021.1
requests>=2.25.1
setuptools>=49.2.1
six>=1.15.0
toml>=0.10.2
urllib3>=1.26.4
virtualenv>=20.4.3
Werkzeug>=1.0.1
wrapt>=1.12.1